+++ regcomp.c
@@ -5863,7 +5863,7 @@ Perl_regprop(pTHX_ const regexp *prog, SV *sv, const regnode *o)
 	}
 
 	if (o->flags & ANYOF_CLASS)
-	    for (i = 0; i < sizeof(anyofs)/sizeof(char*); i++)
+	    for (i = 0; i < (int)(sizeof(anyofs)/sizeof(char*)); i++)
 		if (ANYOF_CLASS_TEST(o,i))
 		    sv_catpv(sv, anyofs[i]);
